requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
flask-wtf>=1.2.0
flask-talisman>=1.1.0
